+++ build_daemon/CHANGELOG.md
@@ -1,5 +1,7 @@
-## 3.0.3-dev
+## 3.1.0-dev
 
+- Add `BuildResults.changedAssets` containing asset URIs changed during a
+  build.
 - Updated the example to use `dart pub` instead of `pub`.
 
 ## 3.0.2